+++ Tickets-services/src/main/java/com/tickets/services/catalogos/EventoService.java
@@ -9,10 +9,10 @@ public interface EventoService {
     
     public Evento actualizaEvento(Evento usuario);
     
-    public Evento bajaEvento(Long id);
+    public Evento bajaEvento(Integer id);
     
     public List<Evento> consultaEventosList();
     
-    public Evento consultaEvento(Long id);
+    public Evento consultaEvento(Integer id);
     
 }
